Put Error{Is,As} matchers behind the build tag
