+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/instana/testify
 
-go 1.13
+go 1.11
 
 require (
 	github.com/davecgh/go-spew v1.1.0
